Predict Recruitment

Predict adjusted recruitment by year using DeCesare et al. (2012) methods. If year is FALSE, predictions are made for a 'typical' year. See bb_predict_calf_cow_ratio() for unadjusted recruitment.

Usage

bb_predict_recruitment(
  recruitment,
  year = TRUE,
  sex_ratio = 0.5,
  conf_level = 0.95,
  estimate = median,
  sig_fig = 3
)

Arguments

recruitment

An object of class 'bboufit_recruitment' (output of bb_fit_recruitment())

year

A flag indicating whether to predict by year.

sex_ratio

A number between 0 and 1 of the proportion of females at birth.

conf_level

A number between 0 and 1 of the confidence level.

estimate

A function to calculate the estimate.

sig_fig

A whole number of the significant figures to round estimates by.

Value

A tibble of the predicted estimates.

References

DeCesare, Nicholas J., Mark Hebblewhite, Mark Bradley, Kirby G. Smith, David Hervieux, and Lalenia Neufeld. 2012 “Estimating Ungulate Recruitment and Growth Rates Using Age Ratios.” The Journal of Wildlife Management 76 (1): 144–53 https://doi.org/10.1002/jwmg.244.

See also